:sparkles: 게시글 작성

+++ backend/src/main/java/com/nts/notice/db/repository/CommentRepository.java
@@ -3,7 +3,7 @@ package com.nts.notice.db.repository;
 import com.nts.notice.db.entity.Comment;
 
 public interface CommentRepository {
-    Comment findbyId(int commentId);
+    Comment findbyId(long commentId);
     void save(Comment comment);
     void delete(Comment comment);
 }
