+++ backend/src/main/java/com/nts/notice/db/repository/FavoriteRepository.java
@@ -3,7 +3,7 @@ package com.nts.notice.db.repository;
 import com.nts.notice.db.entity.Favorite;
 
 public interface FavoriteRepository {
-    Favorite findById(int favoriteId);
+    Favorite findById(long favoriteId);
     void save(Favorite favorite);
     void delete(Favorite favorite);
 }
