+++ backend/src/main/java/com/nts/notice/db/repository/TagRepository.java
@@ -1,10 +1,9 @@
 package com.nts.notice.db.repository;
 
-import com.nts.notice.db.entity.Notice;
 import com.nts.notice.db.entity.Tag;
 
 public interface TagRepository {
-    Tag findById(int tagId);
+    Tag findById(long tagId);
     void save(Tag tag);
     void delete(Tag tag);
 }
